+++ modules/terraform-zscc-gwlb-aws/variables.tf
@@ -70,6 +70,12 @@ variable "cc_instance_size" {
   }
 }
 
+variable "asg_enabled" {
+  type        = bool
+  description = "Determines whether to set gwlb target group target_type to 'instance' or 'ip'. If set to true, ASG uses 'instance' and no aws_lb_target_group_attachment resources need to be created"
+  default     = false
+}
+
 variable "gwlb_name" {
   type        = string
   description = "GWLB resource and tag name"